Implements a readiness endpoint (#82)
* Implements a readiness endpoint

When this converter-application is deployed to a Kubernetes (or Kubernetes-like) environment, a readiness endpoint can be crucial for ensuring that the application has started and is ready to accept new connections.

The same endpoint can also be used in other contexts, like e.g. for testing that another application is able to reach the converter-application (used like a ping endpoint basically).

This new /ready endpoint checks that the officeManager in the converter service is running and can execute tasks, and responds with 200 OK if so. Otherwise it responds with 503 Service Unavailable

* Updates docs for the readiness endpoint